// web/src/lib/services/breadthConstituents.generated.ts
export const BANK_NIFTY = [
  {
    "symbol": "AUBANK",
    "name": "AU Small Finance Bank Ltd.",
    "isin": "INE949L01017",
    "instrumentKey": "NSE_EQ|INE949L01017"
  },
  {
    "symbol": "AXISBANK",
    "name": "Axis Bank Ltd.",
    "isin": "INE238A01034",
    "instrumentKey": "NSE_EQ|INE238A01034"
  },
  {
    "symbol": "BANKBARODA",
    "name": "Bank of Baroda",
    "isin": "INE028A01039",
    "instrumentKey": "NSE_EQ|INE028A01039"
  },
  {
    "symbol": "CANBK",
    "name": "Canara Bank",
    "isin": "INE476A01022",
    "instrumentKey": "NSE_EQ|INE476A01022"
  },
  {
    "symbol": "FEDERALBNK",
    "name": "Federal Bank Ltd.",
    "isin": "INE171A01029",
    "instrumentKey": "NSE_EQ|INE171A01029"
  },
  {
    "symbol": "HDFCBANK",
    "name": "HDFC Bank Ltd.",
    "isin": "INE040A01034",
    "instrumentKey": "NSE_EQ|INE040A01034"
  },
  {
    "symbol": "ICICIBANK",
    "name": "ICICI Bank Ltd.",
    "isin": "INE090A01021",
    "instrumentKey": "NSE_EQ|INE090A01021"
  },
  {
    "symbol": "IDFCFIRSTB",
    "name": "IDFC First Bank Ltd.",
    "isin": "INE092T01019",
    "instrumentKey": "NSE_EQ|INE092T01019"
  },
  {
    "symbol": "INDUSINDBK",
    "name": "IndusInd Bank Ltd.",
    "isin": "INE095A01012",
    "instrumentKey": "NSE_EQ|INE095A01012"
  },
  {
    "symbol": "KOTAKBANK",
    "name": "Kotak Mahindra Bank Ltd.",
    "isin": "INE237A01036",
    "instrumentKey": "NSE_EQ|INE237A01036"
  },
  {
    "symbol": "PNB",
    "name": "Punjab National Bank",
    "isin": "INE160A01022",
    "instrumentKey": "NSE_EQ|INE160A01022"
  },
  {
    "symbol": "SBIN",
    "name": "State Bank of India",
    "isin": "INE062A01020",
    "instrumentKey": "NSE_EQ|INE062A01020"
  },
  {
    "symbol": "UNIONBANK",
    "name": "Union Bank of India",
    "isin": "INE692A01016",
    "instrumentKey": "NSE_EQ|INE692A01016"
  },
  {
    "symbol": "YESBANK",
    "name": "Yes Bank Ltd.",
    "isin": "INE528G01035",
    "instrumentKey": "NSE_EQ|INE528G01035"
  }
] as const;